+++ core/src/mast/serialization/basic_blocks.rs
@@ -30,7 +30,7 @@ impl BasicBlockDataBuilder {
     pub fn encode_basic_block(&mut self, basic_block: &BasicBlockNode) -> NodeDataOffset {
         let ops_offset = self.node_data.len() as NodeDataOffset;
 
-        let operations: Vec<Operation> = basic_block.operations().copied().collect();
+        let operations: Vec<Operation> = basic_block.raw_operations().copied().collect();
         operations.write_into(&mut self.node_data);
 
         ops_offset